Fix Cleanup workflow
Make sure the bpfd-operator owns the bpfd-configmap via a finalizer.

Cleanup the deletion workflow so that the bpfd-daemon is cleaned up
once the configmap is deleted.
