+++ bpfd-operator/config/rbac/bpfd-agent/role.yaml
@@ -42,6 +42,12 @@ rules:
 - apiGroups:
   - bpfd.io
   resources:
+  - tcprograms/finalizers
+  verbs:
+  - update
+- apiGroups:
+  - bpfd.io
+  resources:
   - tracepointprograms
   verbs:
   - get
@@ -50,11 +56,23 @@ rules:
 - apiGroups:
   - bpfd.io
   resources:
+  - tracepointprograms/finalizers
+  verbs:
+  - update
+- apiGroups:
+  - bpfd.io
+  resources:
   - xdpprograms
   verbs:
   - get
   - list
   - watch
+- apiGroups:
+  - bpfd.io
+  resources:
+  - xdpprograms/finalizers
+  verbs:
+  - update
 - apiGroups:
   - ""
   resources:
