+++ bpfd-operator/controllers/bpfd-operator/configmap.go
@@ -18,12 +18,15 @@ package bpfdoperator
 
 import (
 	"context"
+	"io"
+	"os"
 	"reflect"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/kubernetes/scheme"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
@@ -32,12 +35,12 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	operatorinternal "github.com/bpfd-dev/bpfd/bpfd-operator/controllers/bpfd-operator/internal"
 	"github.com/bpfd-dev/bpfd/bpfd-operator/internal"
 )
 
 // +kubebuilder:rbac:groups=apps,resources=daemonsets,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=core,resources=configmaps,verbs=get;list;watch;create
+// +kubebuilder:rbac:groups=bpfd.io,resources=configmaps/finalizers,verbs=update
 
 type BpfdConfigReconciler struct {
 	ReconcilerCommon
@@ -69,7 +72,14 @@ func (r *BpfdConfigReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 			return ctrl.Result{}, nil
 		}
 	} else {
-		return r.ReconcileBpfdConfig(ctx, req, bpfdConfig)
+		if updated := controllerutil.AddFinalizer(bpfdConfig, "bpfd.io.operator/finalizer"); updated {
+			if err := r.Update(ctx, bpfdConfig); err != nil {
+				r.Logger.Error(err, "failed adding bpfd-operator finalizer to bpfd config")
+				return ctrl.Result{Requeue: true, RequeueAfter: retryDurationOperator}, nil
+			}
+		} else {
+			return r.ReconcileBpfdConfig(ctx, req, bpfdConfig)
+		}
 	}
 
 	return ctrl.Result{}, nil
@@ -77,7 +87,7 @@ func (r *BpfdConfigReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 
 func (r *BpfdConfigReconciler) ReconcileBpfdConfig(ctx context.Context, req ctrl.Request, bpfdConfig *corev1.ConfigMap) (ctrl.Result, error) {
 	bpfdDeployment := &appsv1.DaemonSet{}
-	staticBpfdDeployment := operatorinternal.LoadAndConfigureBpfdDs(bpfdConfig)
+	staticBpfdDeployment := LoadAndConfigureBpfdDs(bpfdConfig)
 	r.Logger.V(1).Info("StaticBpfdDeployment is", "DS", staticBpfdDeployment)
 
 	if err := r.Get(ctx, types.NamespacedName{Namespace: bpfdConfig.Data["bpfd.namespace"], Name: internal.BpfdDsName}, bpfdDeployment); err != nil {
@@ -96,6 +106,7 @@ func (r *BpfdConfigReconciler) ReconcileBpfdConfig(ctx context.Context, req ctrl
 	}
 
 	if !bpfdConfig.DeletionTimestamp.IsZero() {
+		r.Logger.Info("Deleting bpfd daemon and config")
 		controllerutil.RemoveFinalizer(bpfdDeployment, bpfdOperatorFinalizer)
 
 		err := r.Update(ctx, bpfdDeployment)
@@ -108,7 +119,14 @@ func (r *BpfdConfigReconciler) ReconcileBpfdConfig(ctx context.Context, req ctrl
 			r.Logger.Error(err, "failed deleting bpfd DS")
 			return ctrl.Result{Requeue: true, RequeueAfter: retryDurationOperator}, nil
 		}
-		
+
+		controllerutil.RemoveFinalizer(bpfdConfig, bpfdOperatorFinalizer)
+		err = r.Update(ctx, bpfdConfig)
+		if err != nil {
+			r.Logger.Error(err, "failed removing bpfd-operator finalizer from bpfd config")
+			return ctrl.Result{Requeue: true, RequeueAfter: retryDurationOperator}, nil
+		}
+
 		return ctrl.Result{}, nil
 	}
 
@@ -159,4 +177,41 @@ func bpfdConfigPredicate() predicate.Funcs {
 			return e.Object.GetName() == internal.BpfdConfigName
 		},
 	}
+}
+
+func LoadAndConfigureBpfdDs(config *corev1.ConfigMap) *appsv1.DaemonSet {
+	// Load static bpfd deployment from disk
+	file, err := os.Open(internal.BpfdDaemonManifestPath)
+	if err != nil {
+		panic(err)
+	}
+
+	b, err := io.ReadAll(file)
+	if err != nil {
+		panic(err)
+	}
+
+	decode := scheme.Codecs.UniversalDeserializer().Decode
+	obj, _, _ := decode(b, nil, nil)
+
+	staticBpfdDeployment := obj.(*appsv1.DaemonSet)
+
+	// Runtime Configurable fields
+	bpfdNamespace := config.Data["bpfd.namespace"]
+	bpfdImage := config.Data["bpfd.image"]
+	bpfdAgentImage := config.Data["bpfd.agent.image"]
+	bpfdLogLevel := config.Data["bpfd.log.level"]
+
+	// Annotate the log level on the ds so we get automatic restarts on changes.
+	if staticBpfdDeployment.Spec.Template.ObjectMeta.Annotations == nil {
+		staticBpfdDeployment.Spec.Template.ObjectMeta.Annotations = make(map[string]string)
+	}
+	staticBpfdDeployment.Spec.Template.ObjectMeta.Annotations["bpfd.io.bpfd.loglevel"] = bpfdLogLevel
+	staticBpfdDeployment.Name = "bpfd-daemon"
+	staticBpfdDeployment.Namespace = bpfdNamespace
+	staticBpfdDeployment.Spec.Template.Spec.Containers[0].Image = bpfdImage
+	staticBpfdDeployment.Spec.Template.Spec.Containers[1].Image = bpfdAgentImage
+	controllerutil.AddFinalizer(staticBpfdDeployment, "bpfd.io.operator/finalizer")
+
+	return staticBpfdDeployment
 }
